For ENG-5803, finally get Paul's suggestion correct

+++ src/ee/storage/table.h
@@ -394,7 +394,7 @@ protected:
         if (m_tuplesPinnedByUndo != 0) {
             return false;
         }
-        return allocatedTupleCount() - activeTupleCount() > std::max((m_tuplesPerBlock * 3), allocatedTupleCount * (100 - m_compactionThreshold / 100));  /* using the integer percentage */
+        return allocatedTupleCount() - activeTupleCount() > std::max(static_cast<int64_t>((m_tuplesPerBlock * 3)), (allocatedTupleCount() * (100 - m_compactionThreshold)) / 100);  /* using the integer percentage */
     }
 
     void initializeWithColumns(TupleSchema *schema, const std::vector<std::string> &columnNames, bool ownsTupleSchema, int32_t compactionThreshold = 95);
